requirements.txt
django==4.0.3
djangorestframework==3.13.1
psycopg2==2.9.3
urllib3==1.26.8
lxml==4.8.0